move transform tests to their own file
